[FLINK-12233][hive] Support table related operations in HiveCatalog
This PR introduced HiveCatalogTable and implemented table related catalog APIs in HiveCatalog.

This closes #8353.
